backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
mistralai>=1.0.0  # Required for Mistral AI integration